First commit
The Eveltic Cookie Consent Bundle is a Symfony package to implement and manage user cookie consents on websites, thus complying fully with the newest 2024 AEPD regulations. It provides an easy-to-use interface to configure, collect, and store the preferences of user consent.

Different types of cookie settings can be configured for technical, analytical, preference, and advertisement cookies. It features easy-to-use templates, multilingual support, and is easy to implement in Symfony applications to ensure that your website is within the bounds of the law.

All operations are managed through ajax requests without the need of any refresh.

Features:
- Manage cookie consent categories: technical, preferences, analytics, and ads (you can configure all categories adding or removing according to your needs).
- Configure expiration period for cookies.
- Support for light, dark, and auto theme modes.
- Command line tools for archiving cookie consent data.
- Integration with Doctrine ORM.
- Fully featured twig display.
- Twig and backend functions to check if consent is accepted.
- Cookie versioning to force the user to reconsent.